+++ med_core/src/app/mod.rs
@@ -1,4 +1,5 @@
 pub mod core;
 pub mod csv;
 pub mod json;
+pub mod processor;
 pub mod worker;
